requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pyjwt==2.10.0
cryptography==43.0.3
pdfplumber==0.11.4
openpyxl==3.1.5
pydantic==2.10.2
pydantic-settings==2.6.1
httpx==0.28.0
